fix: centralize relative time formatting logic
Moved relative time formatting logic from BubbleModel and AppNotifyItem
to a common static method in NotifyEntity to eliminate code duplication
and fix incorrect time display issues. Previously, BubbleModel used
a simple minutes calculation while AppNotifyItem used ICU library for
locale-aware formatting, causing inconsistent time display between
notification bubble and center. Now both components use the same unified
formatting logic.

The changes include:
1. Added static method NotifyEntity::formatRelativeTime() that handles
all relative time formatting using ICU library
2. Removed QDateTime include from bubblemodel.cpp as it's no longer
needed
3. Simplified updateBubbleTimeTip() in BubbleModel to use the new common
method
4. Simplified updateTime() in AppNotifyItem to use the new common method
with fallback to "Just now"
5. Updated copyright years to 2024-2026 in modified files

Log: Fixed inconsistent time display between notification bubble and
center

Influence:
1. Test notification time display in both bubble and center views
2. Verify time formatting for different time intervals (just now,
minutes, hours, yesterday, days, weeks)
3. Check locale-specific formatting works correctly
4. Verify time updates dynamically as notifications age
5. Test with notifications from different time periods

fix: 集中相对时间格式化逻辑

将相对时间格式化逻辑从 BubbleModel 和 AppNotifyItem 移动到 NotifyEntity
的公共静态方法中，以消除代码重复并修复时间显示错误问题。之前，
BubbleModel 使用简单的分钟计算，而 AppNotifyItem 使用 ICU 库进行本地化感
知的格式化，导致通知气泡和中心之间的时间显示不一致。现在两个组件都使用相
同的统一格式化逻辑。

变更包括：
1. 添加静态方法 NotifyEntity::formatRelativeTime()，使用 ICU 库处理所有
相对时间格式化
2. 从 bubblemodel.cpp 中移除不再需要的 QDateTime 包含
3. 简化 BubbleModel 中的 updateBubbleTimeTip() 以使用新的公共方法
4. 简化 AppNotifyItem 中的 updateTime() 以使用新的公共方法，并回退到"刚
刚"
5. 在修改的文件中更新版权年份为 2024-2026

Log: 修复通知气泡和中心之间时间显示不一致的问题

Influence:
1. 测试通知气泡和中心视图中的时间显示
2. 验证不同时间间隔（刚刚、分钟、小时、昨天、天数、周数）的时间格式化
3. 检查本地化特定的格式化是否正确工作
4. 验证时间随通知老化而动态更新
5. 测试来自不同时间段的通知

PMS: BUG-355185
